+++ CHANGES.rst
@@ -7,6 +7,8 @@
 Untagged
 
 - src: Remove CLI restriction on max random-seed (unnecessary since Numpy 1.17).
+- src: Extend ``app`` output with logical commutations and custom values,
+  including possibility to override via ``model.DecodeResult``.
 - docs: Add change list.
 
 1.0b5